Move the image-related models into a separate namespace
