+++ common/internal_model/src/main/scala/uk/ac/wellcome/models/Implicits.scala
@@ -5,6 +5,13 @@ import io.circe._
 import uk.ac.wellcome.models.matcher.MatcherResult
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.models.work.internal._
+import weco.catalogue.internal_model.image.{
+  Image,
+  ImageData,
+  ImageSource,
+  ImageState,
+  SourceWorks
+}
 
 object Implicits {
