+++ common/internal_model/src/main/scala/weco/catalogue/internal_model/image/ImageSource.scala
@@ -1,6 +1,7 @@
-package uk.ac.wellcome.models.work.internal
+package weco.catalogue.internal_model.image
 
 import uk.ac.wellcome.models.work.internal.IdState.Identified
+import uk.ac.wellcome.models.work.internal._
 
 sealed trait ImageSource {
   val id: Identified
