+++ common/internal_model/src/test/scala/uk/ac/wellcome/models/work/generators/MetsWorkGenerators.scala
@@ -1,6 +1,7 @@
 package uk.ac.wellcome.models.work.generators
 
 import uk.ac.wellcome.models.work.internal.{Work, WorkState}
+import weco.catalogue.internal_model.generators.ImageGenerators
 
 trait MetsWorkGenerators extends WorkGenerators with ImageGenerators {
   def metsIdentifiedWork(): Work.Visible[WorkState.Identified] =
