+++ common/internal_model/src/test/scala/weco/catalogue/internal_model/generators/VectorGenerators.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models.work.generators
+package weco.catalogue.internal_model.generators
 
 import scala.util.Random
