+++ pipeline/inferrer/inference_manager/src/main/scala/uk/ac/wellcome/platform/inference_manager/Main.scala
@@ -29,8 +29,9 @@ import uk.ac.wellcome.pipeline_storage.typesafe.{
   PipelineStorageStreamBuilder
 }
 import uk.ac.wellcome.elasticsearch.typesafe.ElasticBuilder
-import uk.ac.wellcome.models.work.internal.{Image, ImageState}
-import ImageState.{Augmented, Initial}
+import uk.ac.wellcome.models.work.internal.ImageState
+import weco.catalogue.internal_model.image.Image
+import weco.catalogue.internal_model.image.ImageState.{Augmented, Initial}
 
 object Main extends WellcomeTypesafeApp {
   runWithConfig { config: Config =>
