+++ pipeline/inferrer/inference_manager/src/main/scala/uk/ac/wellcome/platform/inference_manager/adapters/FeatureVectorInferrerAdapter.scala
@@ -6,11 +6,11 @@ import java.util.Base64
 import akka.http.scaladsl.model.{HttpMethods, HttpRequest, Uri}
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
-import uk.ac.wellcome.models.work.internal.InferredData
 import uk.ac.wellcome.platform.inference_manager.models.{
   DownloadedImage,
   FeatureVectorInferrerResponse
 }
+import weco.catalogue.internal_model.image.InferredData
 
 import scala.util.{Success, Try}
