+++ pipeline/inferrer/inference_manager/src/main/scala/uk/ac/wellcome/platform/inference_manager/adapters/PaletteInferrerAdapter.scala
@@ -3,12 +3,12 @@ package uk.ac.wellcome.platform.inference_manager.adapters
 import akka.http.scaladsl.model.{HttpMethods, HttpRequest, Uri}
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
-import uk.ac.wellcome.models.work.internal.InferredData
 import uk.ac.wellcome.platform.inference_manager.models.{
   DownloadedImage,
   HashParams,
   PaletteInferrerResponse
 }
+import weco.catalogue.internal_model.image.InferredData
 
 class PaletteInferrerAdapter(host: String, port: Int) extends InferrerAdapter {
   type Response = PaletteInferrerResponse
