+++ pipeline/inferrer/inference_manager/src/main/scala/uk/ac/wellcome/platform/inference_manager/services/package.scala
@@ -3,6 +3,7 @@ package uk.ac.wellcome.platform.inference_manager
 import akka.http.scaladsl.model.{HttpRequest, HttpResponse}
 import akka.stream.scaladsl.{Flow, FlowWithContext}
 import uk.ac.wellcome.models.work.internal._
+import weco.catalogue.internal_model.image.{Image, ImageState}
 
 import scala.util.Try
