+++ pipeline/inferrer/inference_manager/src/test/scala/uk/ac/wellcome/platform/inference_manager/integration/ManagerInferrerIntegrationTest.scala
@@ -15,13 +15,7 @@ import software.amazon.awssdk.services.sqs.model.Message
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.messaging.fixtures.SQS.QueuePair
 import uk.ac.wellcome.messaging.memory.MemoryMessageSender
-import uk.ac.wellcome.models.work.generators.ImageGenerators
-import uk.ac.wellcome.models.work.internal.{
-  Image,
-  ImageState,
-  InferredData,
-  LocationType
-}
+import uk.ac.wellcome.models.work.internal.LocationType
 import uk.ac.wellcome.platform.inference_manager.adapters.{
   FeatureVectorInferrerAdapter,
   InferrerAdapter,
@@ -33,8 +27,10 @@ import uk.ac.wellcome.platform.inference_manager.services.{
   DefaultFileWriter,
   MergedIdentifiedImage
 }
-import ImageState.{Augmented, Initial}
+import weco.catalogue.internal_model.image.ImageState.{Augmented, Initial}
 import akka.http.scaladsl.model.Uri
+import weco.catalogue.internal_model.generators.ImageGenerators
+import weco.catalogue.internal_model.image.{Image, InferredData}
 
 class ManagerInferrerIntegrationTest
     extends AnyFunSpec
