+++ pipeline/inferrer/inference_manager/src/test/scala/uk/ac/wellcome/platform/inference_manager/services/InferenceManagerWorkerServiceTest.scala
@@ -10,8 +10,7 @@ import software.amazon.awssdk.services.sqs.model.Message
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.messaging.fixtures.SQS.QueuePair
 import uk.ac.wellcome.messaging.memory.MemoryMessageSender
-import uk.ac.wellcome.models.work.internal.{Image, ImageState, InferredData}
-import uk.ac.wellcome.models.work.generators.ImageGenerators
+import uk.ac.wellcome.models.work.internal.InferredData
 import uk.ac.wellcome.platform.inference_manager.adapters.{
   FeatureVectorInferrerAdapter,
   InferrerAdapter,
@@ -25,7 +24,9 @@ import uk.ac.wellcome.platform.inference_manager.fixtures.{
   Responses
 }
 import uk.ac.wellcome.platform.inference_manager.models.DownloadedImage
-import ImageState.{Augmented, Initial}
+import weco.catalogue.internal_model.generators.ImageGenerators
+import weco.catalogue.internal_model.image.{Image, ImageState, InferredData}
+import weco.catalogue.internal_model.image.ImageState.{Augmented, Initial}
 
 class InferenceManagerWorkerServiceTest
     extends AnyFunSpec
