+++ pipeline/ingestor/ingestor_images/src/main/scala/uk/ac/wellcome/platform/ingestor/images/ImageTransformer.scala
@@ -1,7 +1,8 @@
 package uk.ac.wellcome.platform.ingestor.images
 
-import uk.ac.wellcome.models.work.internal.{Image, ImageState}
-import ImageState.{Augmented, Indexed}
+import uk.ac.wellcome.models.work.internal.ImageState
+import weco.catalogue.internal_model.image.Image
+import weco.catalogue.internal_model.image.ImageState.{Augmented, Indexed}
 
 object ImageTransformer {
   val deriveData: Image[Augmented] => Image[Indexed] =
