+++ pipeline/ingestor/ingestor_images/src/test/scala/uk/ac/wellcome/platform/ingestor/images/services/ImagesIndexerTest.scala
@@ -4,12 +4,13 @@ import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.models.index.{IndexFixtures, IndexedImageIndexConfig}
-import uk.ac.wellcome.models.work.generators.ImageGenerators
-import uk.ac.wellcome.models.work.internal.{Image, ImageState}
+import uk.ac.wellcome.models.work.internal.ImageState
 import uk.ac.wellcome.pipeline_storage.ElasticIndexer
 import uk.ac.wellcome.pipeline_storage.Indexable.imageIndexable
 import uk.ac.wellcome.models.Implicits._
 import uk.ac.wellcome.pipeline_storage.fixtures.ElasticIndexerFixtures
+import weco.catalogue.internal_model.generators.ImageGenerators
+import weco.catalogue.internal_model.image.{Image, ImageState}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
