+++ pipeline/merger/src/main/scala/uk/ac/wellcome/platform/merger/models/ImageDataWithSource.scala
@@ -1,6 +1,7 @@
 package uk.ac.wellcome.platform.merger.models
 
-import uk.ac.wellcome.models.work.internal.{IdState, ImageData, ImageSource}
+import uk.ac.wellcome.models.work.internal.IdState
+import weco.catalogue.internal_model.image.{ImageData, ImageSource}
 
 case class ImageDataWithSource(
   imageData: ImageData[IdState.Identified],
